refactor(api): split admin.js controller into per-responsibility modules (#599)
admin.js had grown to 835 lines (audit finding C1). Split it by
responsibility into focused routers under routes/admin/:

  curatedData.js — /api/admin/export|import/curated (+ tableExists / importCuratedTag helpers)
  riskConfig.js  — /api/admin/risk-profile + /api/admin/classifiers
  maintenance.js — /api/admin/clean-database + /api/admin/history-retention (read/write/prune)
  dashboard.js   — /api/admin/dashboard-stats + /api/admin/dashboard-timeseries
  settings.js    — /api/admin/features/toggle + /api/admin/auth-settings

admin.js is now a barrel that composes the five routers, so app.js and
every public path are unchanged. Pure code move — the only adjustment is
the curated-import handler's dynamic-import paths, shifted one level
deeper now that it lives under admin/.

Grandfathered complexity entries relocated (no increase): cyclomatic
[30] -> curatedData; cognitive [46,26,22,22,18] -> curatedData
[46,26,22,22] + maintenance [18].

Added admin.mock.test.js covering the USE_SQL-off "database not
configured" guards the split moved — diff coverage of the changed lines
is 87% (unit-only), clearing the 80% gate. Full API suite green (1752
tests); coverage above every floor; both complexity gates and the
OpenAPI-drift guard pass.
