+++ .ci/cognitive-baseline.json
@@ -60,11 +60,13 @@
       19,
       16
     ],
-    "app/api/src/routes/admin.js": [
+    "app/api/src/routes/admin/curatedData.js": [
       46,
       26,
       22,
-      22,
+      22
+    ],
+    "app/api/src/routes/admin/maintenance.js": [
       18
     ],
     "app/api/src/routes/authRoles.js": [
