+++ .ci/complexity-baseline.json
@@ -31,7 +31,7 @@
       24,
       22
     ],
-    "app/api/src/routes/admin.js": [
+    "app/api/src/routes/admin/curatedData.js": [
       30
     ],
     "app/api/src/routes/contexts.js": [
